src/requirements.txt
django~=4.2
django-debug-toolbar~=3.8
gunicorn~=20.1
whitenoise==6.4
